pyproject.toml
[tool.poetry]
name = "img_locator"
version = "0.1.0"
description = ""
authors = ["Jared Bishop <jared.bishop@pinnacol.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
Jinja2 = "^3.1.2"
python-multipart = "^0.0.5"
python-magic = "^0.4.26"
Pillow = "^9.1.1"
pytest = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
